library/serato: Move SeratoPlaylistModel into separate file
