+++ src/library/serato/seratoplaylistmodel.h
@@ -1,0 +1,19 @@
+#pragma once
+// seratoplaylistmodel.h
+// Created 2020-02-15 by Jan Holthuis
+#include "library/baseexternalplaylistmodel.h"
+
+class TrackCollectionManager;
+class BaseExternalPlaylistModel;
+
+class SeratoPlaylistModel : public BaseExternalPlaylistModel {
+  public:
+    SeratoPlaylistModel(QObject* parent,
+            TrackCollectionManager* pTrackCollectionManager,
+            QSharedPointer<BaseTrackCache> trackSource);
+    TrackPointer getTrack(const QModelIndex& index) const override;
+    bool isColumnHiddenByDefault(int column) override;
+
+  protected:
+    void initSortColumnMapping() override;
+};
